adds Footer
