+++ client/food-recipes-app/src/Components/Dashboard.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Dashboard = () => {
+  return (
+    <div>
+      <h1>Dashboard</h1>
+    </div>
+  );
+};
+
+export default Dashboard;
